requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
tenacity==8.5.0
aiohttp>=3.8.0
aiosqlite==0.20.0
pytz==2024.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
flask>=2.3.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
requests>=2.31.0  